Converted ModifyPhotoshoots to be a dialog
