+++ app/src/main/java/com/jamessimshaw/cosplaycompanion/controllers/ShowConventionController.java
@@ -19,6 +19,7 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.jamessimshaw.cosplaycompanion.CosplayCompanionApplication;
 import com.jamessimshaw.cosplaycompanion.R;
 import com.jamessimshaw.cosplaycompanion.adapters.ConYearRecViewAdapter;
+import com.jamessimshaw.cosplaycompanion.fragments.ModifyConventionYearDialogFragment;
 import com.jamessimshaw.cosplaycompanion.models.ConventionYear;
 import com.jamessimshaw.cosplaycompanion.presenters.ListConventionYearsPresenterImpl;
 import com.jamessimshaw.cosplaycompanion.views.ListConventionYearsView;
