+++ app/src/main/res/layout/fragment_new_photoshoot.xml
@@ -4,7 +4,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             app:layout_behavior="@string/appbar_scrolling_view_behavior"
-            tools:context="com.jamessimshaw.cosplaycompanion.controllers.ModifyPhotoshootController">
+            tools:context="com.jamessimshaw.cosplaycompanion.controllers.ModifyPhotoshootFragment">
 
     <android.support.constraint.ConstraintLayout android:layout_width="match_parent"
                                                  android:layout_height="wrap_content">
@@ -66,6 +66,19 @@
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toBottomOf="@id/seriesEditText"/>
+
+        <TextView
+            android:id="@+id/dialogSubmitTextView"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="8dp"
+            android:layout_marginBottom="8dp"
+            android:layout_marginEnd="16dp"
+            android:layout_marginRight="16dp"
+            app:layout_constraintTop_toBottomOf="@id/descriptionEditText"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:text="Submit"/>
 
     </android.support.constraint.ConstraintLayout>
 </ScrollView>
